+++ example/lib/features/out/out_page.dart
@@ -7,7 +7,7 @@ class OutPage extends StatelessWidget {
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
-        title: Text('out page'),
+        title: const Text('out page'),
       ),
     );
   }